third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGImageElement-dom-y-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' attribute of the SVGImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageElement.getAttribute('y') is "-190"
PASS imageElement.getAttribute('y') is "0"
PASS successfullyParsed is true

TEST COMPLETE
